requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
biopython>=1.81
requests>=2.28.0
scipy>=1.10.0
tqdm>=4.65.0
logomaker>=0.8
scikit-learn>=1.2.0
rdkit-pypi>=2022.9.0
